requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
playwright>=1.40.0
lxml>=4.9.0
urllib3>=2.0.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.2
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.7.0
json5>=0.9.0
webdriver-manager>=4.0.0
fake-useragent>=1.4.0
dnspython>=2.4.0
subprocess32>=3.5.4
requests-html>=0.10.0 
playwright